#209: Refine Tomcat installation to prepare
* Extract tomcat installation/definition to new source files
* Add jsh/tools/install plugin to test suite
